# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic AI Agricultural Crop Planning & Precision Farming System",
    "summary": "Agentic multi-agent crop planning and precision farming API.",
    "description": "\nMulti-agent decision support for crop planning and precision farming.\n\nThe workflow is orchestrated with **LangGraph** over twelve specialised agents\n(farm profile, soil & nutrient, weather & climate, agricultural knowledge RAG,\ncrop suitability, ML forecasting, irrigation, crop-risk review, activity\nplanning and advisory). Every agent emits structured evidence, and every\nphysical action (irrigation in particular) requires **human approval** before it\ncan leave the `planned` state.\n",
    "contact": {
      "name": "Precision Farming Platform"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List generated reports",
        "operationId": "list_reports_api_v1_reports_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field Id"
            }
          },
          {
            "name": "farm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportOut"
                  },
                  "title": "Response List Reports Api V1 Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate a PDF report for a field",
        "operationId": "create_report_api_v1_reports_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report metadata",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reports"
        ],
        "summary": "Delete a report",
        "operationId": "delete_report_api_v1_reports__report_id__delete",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download the PDF",
        "operationId": "download_api_v1_reports__report_id__download_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/documents": {
      "get": {
        "tags": [
          "knowledge base"
        ],
        "summary": "Indexed reference documents",
        "operationId": "documents_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferenceDocumentOut"
                  },
                  "title": "Response Documents Api V1 Knowledge Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/search": {
      "post": {
        "tags": [
          "knowledge base"
        ],
        "summary": "Semantic search over the corpus",
        "operationId": "search_api_v1_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents/{doc_key}": {
      "get": {
        "tags": [
          "knowledge base"
        ],
        "summary": "Citations for one known document",
        "operationId": "document_api_v1_knowledge_documents__doc_key__get",
        "parameters": [
          {
            "name": "doc_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Document Api V1 Knowledge Documents  Doc Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "doc_key": {
            "type": "string",
            "title": "Doc Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "organisation": {
            "type": "string",
            "title": "Organisation"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "doc_key",
          "title",
          "category",
          "organisation",
          "source_url",
          "score",
          "text"
        ],
        "title": "KnowledgeChunk"
      },
      "KnowledgeQuery": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 4
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "KnowledgeQuery"
      },
      "KnowledgeQueryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "index_backend": {
            "type": "string",
            "title": "Index Backend"
          },
          "corpus_documents": {
            "type": "integer",
            "title": "Corpus Documents"
          },
          "corpus_chunks": {
            "type": "integer",
            "title": "Corpus Chunks"
          },
          "retrieval_mode": {
            "type": "string",
            "title": "Retrieval Mode"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeChunk"
            },
            "type": "array",
            "title": "Chunks"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "type": "array",
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "query",
          "index_backend",
          "corpus_documents",
          "corpus_chunks",
          "retrieval_mode",
          "chunks",
          "references"
        ],
        "title": "KnowledgeQueryResponse"
      },
...
      "ReferenceDocumentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "doc_key": {
            "type": "string",
            "title": "Doc Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "organisation": {
            "type": "string",
            "title": "Organisation"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doc_key",
          "title",
          "category",
          "organisation",
          "source_url",
          "region",
          "chunk_count",
          "checksum",
          "is_active"
        ],
        "title": "ReferenceDocumentOut"
      },
      "ReportCreate": {
        "properties": {
          "workflow_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Run Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Defaults to a generated title"
          },
          "include_evidence": {
            "type": "boolean",
            "title": "Include Evidence",
            "default": true
          }
        },
        "type": "object",
        "title": "ReportCreate"
      },
      "ReportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "field_id": {
            "type": "integer",
            "title": "Field Id"
          },
          "workflow_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Run Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "section_summary": {
            "type": "object",
            "title": "Section Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "farm_id",
          "field_id",
          "workflow_run_id",
          "title",
          "status",
          "file_name",
          "size_bytes",
          "page_count",
          "section_summary",
          "error",
          "created_at"
        ],
        "title": "ReportOut"
      },
...
      "SourceReference": {
        "properties": {
          "doc_key": {
            "type": "string",
            "title": "Doc Key",
            "examples": [
              "crops.rice"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "organisation": {
            "type": "string",
            "title": "Organisation",
            "default": ""
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          }
        },
        "type": "object",
        "required": [
          "doc_key",
          "title"
        ],
        "title": "SourceReference",
        "description": "Citation for retrieved reference material or an upstream data provider."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}